+++ VERSION_HISTORY.md
@@ -7,4 +7,19 @@ Historial de Revisiones
 	>
 	>Eduardo A. Farías Reyes
 
+* 26-11-2010 v0.2 
+	>
+	>Eduardo A. Farías Reyes
+	>
+	>Aplicado framework Bashinator, exportación de tablas completa.
+	>Bugs #17 y #18
+
+* 05-12-2010 v0.3 Funcional
+	>
+	>Eduardo A. Farías Reyes
+	>
+	>Script plenamente funcional
+	>Exporta registros desde access hacia mysql
+	>Bugs #17 y #18 solucionados
+